Change date menu from protocol to concrete class
Also added customizable date formatter
